+++ src/main/java/com/itesm/panoptimize/service/StatusService.java
@@ -1,10 +1,12 @@
 package com.itesm.panoptimize.service;
 
+import com.itesm.panoptimize.dto.agent.StatusDTO;
 import software.amazon.awssdk.services.connect.ConnectClient;
 import software.amazon.awssdk.services.connect.model.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import java.util.List;
+
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,19 +25,22 @@ public class StatusService {
      * @param instanceId the ID of the Amazon Connect instance
      * @return a list of active agent statuses
      */
-    public List<AgentStatus> getActiveAgents(String instanceId) {
+    public List<StatusDTO> getActiveAgents(String instanceId) {
         List<String> queueIds = getAllQueueIds(instanceId);
         if (queueIds.isEmpty()) {
             throw new RuntimeException("No queues found for instance: " + instanceId);
         }
 
-        List<CurrentMetric> metrics = List.of(
-                CurrentMetric.builder().name(CurrentMetricName.AGENTS_ONLINE).unit(Unit.COUNT).build(),
-                CurrentMetric.builder().name(CurrentMetricName.AGENTS_AVAILABLE).unit(Unit.COUNT).build(),
-                CurrentMetric.builder().name(CurrentMetricName.AGENTS_AFTER_CONTACT_WORK).unit(Unit.COUNT).build(),
-                CurrentMetric.builder().name(CurrentMetricName.AGENTS_ON_CALL).unit(Unit.COUNT).build()
-
+        List<CurrentMetricName> metricNames = List.of(
+                CurrentMetricName.AGENTS_ONLINE,
+                CurrentMetricName.AGENTS_AVAILABLE,
+                CurrentMetricName.AGENTS_AFTER_CONTACT_WORK,
+                CurrentMetricName.AGENTS_ON_CONTACT
         );
+
+        List<CurrentMetric> metrics = metricNames.stream()
+                .map(metricName -> CurrentMetric.builder().name(metricName).unit(Unit.COUNT).build())
+                .collect(Collectors.toList());
 
         GetCurrentMetricDataRequest request = GetCurrentMetricDataRequest.builder()
                 .instanceId(instanceId)
@@ -46,11 +51,18 @@ public class StatusService {
         try {
             GetCurrentMetricDataResponse response = connectClient.getCurrentMetricData(request);
 
-            return response.metricResults().stream()
+            Map<String, Double> metricValues = metricNames.stream()
+                    .collect(Collectors.toMap(
+                            CurrentMetricName::toString,
+                            metricName -> 0.0
+                    ));
+
+            response.metricResults().stream()
                     .flatMap(result -> result.collections().stream())
-                    .filter(collection -> collection.metric().name().equals(CurrentMetricName.AGENTS_ONLINE) ||
-                            collection.metric().name().equals(CurrentMetricName.AGENTS_AVAILABLE))
-                    .map(collection -> new AgentStatus(collection.metric().nameAsString(), collection.value()))
+                    .forEach(collection -> metricValues.put(collection.metric().nameAsString(), collection.value()));
+
+            return metricNames.stream()
+                    .map(metricName -> new StatusDTO(metricName.toString(), metricValues.get(metricName.toString())))
                     .collect(Collectors.toList());
         } catch (ConnectException e) {
             // Log the exception and rethrow or handle it as needed
@@ -74,26 +86,5 @@ public class StatusService {
         return listQueuesResponse.queueSummaryList().stream()
                 .map(QueueSummary::id)
                 .collect(Collectors.toList());
-    }
-
-    /**
-     * Inner class representing the status of an agent.
-     */
-    public static class AgentStatus {
-        private final String metricName;
-        private final Double metricValue;
-
-        public AgentStatus(String metricName, Double metricValue) {
-            this.metricName = metricName;
-            this.metricValue = metricValue;
-        }
-
-        public String getMetricName() {
-            return metricName;
-        }
-
-        public Double getMetricValue() {
-            return metricValue;
-        }
     }
 }
